+++ ext-src/IqComponentModel.ts
@@ -13,12 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import * as request from "request";
 import { Uri, window } from "vscode";
-import * as HttpStatus from 'http-status-codes';
 
 import { ComponentEntry, PolicyViolation } from "./ComponentInfoPanel";
 import { ComponentContainer } from "./packages/ComponentContainer";
+import { IqRequestService } from "./IqRequestService";
+import { RequestService } from "./RequestService";
 
 export class IqComponentModel {
     components: Array<ComponentEntry> = [];
@@ -26,9 +26,7 @@ export class IqComponentModel {
       string,
       ComponentEntry
     >();
-  
-    // TODO make these configurable???
-    readonly evaluationPollDelayMs = 2000;
+    requestService: RequestService;
   
     constructor(
       readonly url: string,
@@ -36,7 +34,9 @@ export class IqComponentModel {
       private password: string,
       private applicationPublicId: string,
       private getmaximumEvaluationPollAttempts: number
-    ) {}
+    ) {
+      this.requestService = new IqRequestService(this.url, this.user, this.password, this.getmaximumEvaluationPollAttempts);
+    }
 
     public setPassword(password: string) {
       this.password = password;
@@ -50,7 +50,6 @@ export class IqComponentModel {
     }
   
     public getContent(resource: Uri): Thenable<string> {
-      // TODO get the HTML doc for webview
       return new Promise((c, e) => "my stubbed content entry");
     }
   
@@ -61,7 +60,7 @@ export class IqComponentModel {
   
     private async performIqScan() {
       try {
-        let componentContainer = new ComponentContainer();
+        let componentContainer = new ComponentContainer(this.requestService);
 
         let data: any;
 
@@ -80,7 +79,7 @@ export class IqComponentModel {
         }
   
         console.debug("getting applicationInternalId", this.applicationPublicId);
-        let response: string = await this.getApplicationId(this.applicationPublicId) as string;
+        let response: string = await this.requestService.getApplicationId(this.applicationPublicId) as string;
   
         let appRep = JSON.parse(response);
         console.debug("appRep", appRep);
@@ -88,10 +87,10 @@ export class IqComponentModel {
         let applicationInternalId: string = appRep.applications[0].id;
         console.debug("applicationInternalId", applicationInternalId);
   
-        let resultId = await this.submitToIqForEvaluation(data, applicationInternalId);
+        let resultId = await this.requestService.submitToIqForEvaluation(data, applicationInternalId);
   
         console.debug("report", resultId);
-        let resultDataString = await this.asyncPollForEvaluationResults(applicationInternalId, resultId);
+        let resultDataString = await this.requestService.asyncPollForEvaluationResults(applicationInternalId, resultId);
         let resultData = JSON.parse(resultDataString as string);
   
         console.debug(`Received results from IQ scan:`, resultData);
@@ -111,302 +110,5 @@ export class IqComponentModel {
         window.showErrorMessage("Nexus IQ extension: " + e);
         return;
       }
-  }
-
-  private async getApplicationId(applicationPublicId: string) {
-    console.debug("getApplicationId", applicationPublicId);
-
-    return new Promise((resolve, reject) => {
-      request.get(
-        {
-          method: "GET",
-          url: `${this.url}/api/v2/applications?publicId=${applicationPublicId}`,
-          auth: { user: this.user, pass: this.password }
-        },
-        (err: any, response: any, body: any) => {
-          if (err) {            
-            reject(`Unable to retrieve Application ID: ${err}`);
-            return;
-          }
-          if (response.statusCode != HttpStatus.OK) {            
-            reject(`Unable to retrieve Application ID. Could not communicate with server. Server error: ${response.statusCode}`);
-            return;
-          }
-          resolve(body);
-          return;
-        }
-      );
-    });
-  }
-
-  private async submitToIqForEvaluation(
-    data: any,
-    applicationInternalId: string
-  ): Promise<string> {
-    return new Promise((resolve, reject) => {
-      request.post(
-        {
-          method: "POST",
-          url: `${this.url}/api/v2/evaluation/applications/${applicationInternalId}`,
-          json: data,
-          auth: { user: this.user, pass: this.password }
-        },
-        (err: any, response: any, body: any) => {
-          // console.log(response.statusCode, body);
-          if (err) {
-            reject(`Unable to perform IQ scan: ${err}`);
-            return;
-          }
-          if (response.statusCode != HttpStatus.OK) {
-            reject(`Unable to perform IQ scan: ${body}`);
-            return;
-          }
-
-          let resultId = body.resultId;
-          resolve(resultId);
-          return;
-        }
-      );
-    });
-  }
-
-  private async asyncPollForEvaluationResults(
-    applicationInternalId: string,
-    resultId: string
-  ) {
-    return new Promise((resolve, reject) => {
-      this.pollForEvaluationResults(
-        applicationInternalId,
-        resultId,
-        body => resolve(body),
-        (statusCode, message) =>
-          reject(
-            `Could not fetch evaluation result, code ${statusCode}, message ${message}`
-          )
-      );
-    });
-  }
-
-  private pollForEvaluationResults(
-    applicationInternalId: string,
-    resultId: string,
-    success: (body: string) => any,
-    failed: (statusCode: number, message: string) => any
-  ) {
-    let _this = this;
-    let pollAttempts = 0;
-
-    let successHandler = function(value: string) {
-      success(value);
-    };
-    let errorHandler = function(statusCode: number, message: string) {
-      if (statusCode === HttpStatus.NOT_FOUND) {
-        // report still being worked on, continue to poll
-        pollAttempts += 1;
-        // TODO use the top-level class constant, but references are failing
-        if (pollAttempts >= _this.getmaximumEvaluationPollAttempts) {
-          failed(statusCode, "Poll limit exceeded, try again later");
-        } else {
-          setTimeout(() => {
-            _this.getEvaluationResults(
-              applicationInternalId,
-              resultId,
-              successHandler,
-              errorHandler
-            );
-          }, _this.evaluationPollDelayMs);
-        }
-      } else {
-        failed(statusCode, message);
-      }
-    };
-    this.getEvaluationResults(
-      applicationInternalId,
-      resultId,
-      successHandler,
-      errorHandler
-    );
-  }
-
-  private getEvaluationResults(
-    applicationInternalId: string,
-    resultId: string,
-    resolve: (body: string) => any,
-    reject: (statusCode: number, message: string) => any
-  ) {
-    request.get(
-      {
-        method: "GET",
-        url: `${this.url}/api/v2/evaluation/applications/${applicationInternalId}/results/${resultId}`,
-        auth: { user: this.user, pass: this.password }
-      },
-      (error: any, response: any, body: any) => {
-        if (response && response.statusCode != HttpStatus.OK) {
-          reject(response.statusCode, error);
-          return;
-        }
-        if (error) {
-          reject(response.statusCode, error);
-          return;
-        }
-        resolve(body);
-      }
-    );
-  }
-
-  public async getRemediation(nexusArtifact: any, iqApplicationId: string) {
-    return new Promise((resolve, reject) => {
-      console.debug("begin getRemediation", nexusArtifact);
-      var requestdata = nexusArtifact.component;
-      console.debug("requestdata", requestdata);
-      let url = `${this.url}/api/v2/components/remediation/application/${iqApplicationId}`;
-
-      request.post(
-        {
-          method: "post",
-          json: requestdata,
-          url: url,
-          auth: { user: this.user, pass: this.password }
-        },
-        (err, response, body) => {
-          if (err) {
-            reject(`Unable to retrieve Component details: ${err}`);
-            return;
-          }
-          console.debug("response", response);
-          console.debug("body", body);
-          resolve(body);
-        }
-      );
-    });
-  }
-
-  public async GetCVEDetails(cve: any, nexusArtifact: any) {
-    //, settings) {
-    return new Promise((resolve, reject) => {
-      console.log("begin GetCVEDetails", cve, nexusArtifact);
-      let timestamp = Date.now();
-      let hash = nexusArtifact.components[0].hash;
-      let componentIdentifier = this.encodeComponentIdentifier(
-        nexusArtifact.components[0].componentIdentifier
-      );
-      let vulnerability_source;
-      if (cve.search("sonatype") >= 0) {
-        vulnerability_source = "sonatype";
-      } else {
-        vulnerability_source = "cve";
-      }
-      let url = `${this.url}/rest/vulnerability/details/${vulnerability_source}/${cve}?componentIdentifier=${componentIdentifier}&hash=${hash}&timestamp=${timestamp}`;
-
-      request.get(
-        {
-          method: "GET",
-          url: url,
-          auth: {
-            user: this.user,
-            pass: this.password
-          }
-        },
-        (err, response, body) => {
-          if (err) {
-            reject(`Unable to retrieve CVEData: ${err}`);
-            return;
-          }
-          console.debug("response", response);
-          console.debug("body", body);
-
-          resolve(body);
-        }
-      );
-    });
-  }
-
-  public async getAllVersions(nexusArtifact: any, iqApplicationPublicId: string): Promise<any[]> {
-    if (!nexusArtifact || !nexusArtifact.hash) {
-      return [];
-    }
-    return new Promise<any[]>((resolve, reject) => {
-      let hash = nexusArtifact.hash;
-      let comp = this.encodeComponentIdentifier(
-        nexusArtifact.componentIdentifier
-      );
-      let d = new Date();
-      let timestamp = d.getDate();
-      let matchstate = "exact";
-      let url =
-        `${this.url}/rest/ide/componentDetails/application/` +
-        `${iqApplicationPublicId}/allVersions?` +
-        `componentIdentifier=${comp}&` +
-        `hash=${hash}&matchState=${matchstate}&` +
-        `timestamp=${timestamp}&proprietary=false`;
-
-      request.get(
-        {
-          method: "GET",
-          url: url,
-          auth: {
-            user: this.user,
-            pass: this.password
-          }
-        },
-        (err, response, body) => {
-          if (err) {
-            reject(`Unable to retrieve GetAllVersions: ${err}`);
-            return;
-          }
-          const versionArray = JSON.parse(body) as any[];
-          console.debug("getAllVersions retrieved body", versionArray);
-          resolve(versionArray);
-        }
-      );
-    });
-  }
-
-  public async showSelectedVersion(componentIdentifier: any, version: string) {
-    return new Promise((resolve, reject) => {
-      console.debug("begin showSelectedVersion", componentIdentifier, version);
-      var transmittingComponentIdentifier = { ...componentIdentifier };
-
-      transmittingComponentIdentifier.coordinates = {
-        ...componentIdentifier.coordinates
-      };
-
-      transmittingComponentIdentifier.coordinates.version = version;
-      var detailsRequest = {
-        components: [
-          {
-            hash: null,
-            componentIdentifier: transmittingComponentIdentifier
-          }
-        ]
-      };
-      let url = `${this.url}/api/v2/components/details`;
-
-      request.post(
-        {
-          method: "post",
-          json: detailsRequest,
-          url: url,
-          auth: {
-            user: this.user,
-            pass: this.password
-          }
-        },
-        (err, response, body) => {
-          if (err) {
-            reject(`Unable to retrieve selected version details: ${err}`);
-            return;
-          }
-
-          resolve(body);
-        }
-      );
-    });
-  }
-
-  private encodeComponentIdentifier(componentIdentifier: string) {
-    let actual = encodeURIComponent(JSON.stringify(componentIdentifier));
-    console.log("actual", actual);
-    return actual;
   }
 }
